ccxmx5x: add power_off support
Not that the implementation differs from the one from FSL.
This one is similar to what we have implemented in WinCE.
The procedure comes directly from Dialog.
We dont know why the method used by FSL in their QSB it
doesnt work in our module.
She the comments in the code for additional details.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>