ccxmx5x: add support for HSD101PFW2 display
This commit adds support for a new LCD display (HSD101PFW2).
This display has been tested with the ConnectCore Wi-i.MX53,
connected to its LVDS interface.
Keep in mind that the backlight control needs to be controlled
through the PWM input of the display. In the CCWi-iMX53 JSK
board this is done through the PWM led output (PMIC GPIO15).

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>