ccxmx5x dual video: remove code to make it work
This code skips the execution of some code in the
mxc_ipuv3_fb driver when dual head is used. This code
was causing that the dual head was not working.
This patch was implemented time ago for the wimx51 but
had been lost while merging recent patches from FSL.
It is not clear why we need this. Needs to be investigated.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>