ccxmx53: da9052-tsi, disable window filter.
The driver enables by default a window filter that stops tracking movement
if it detects a pixel difference of 50 pixels between positions. Disabling
this "feature" fixes this Vantive.

This commit allows for the driver to be configured to use the window or
not, with the default being to disable it.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>