ccxmx53: Do not reconfigure IOMUX select input twice for UART0/1
The CTS and RTS pins for UARTS0 and 1 are both set in their respective
select input registers. When both pins are reconfigured, only the last
IOMUX configuration remains.

This change makes the first IOMUX configuration take effect, while the
second IOMUX configuration will affect the pad control but not the select
input control.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>