mxc ldb: fix problem with dual modes
This commit fixes a problem with the dual mode that was
happening in following scenario:

video=displayfb:disabled
video2=displayfb:LCD@LQ121K1LG11
ldb=dual,di=1,ch1_map=JEIDA

Here the image on the LVDS channel 1 was not correctly
represented because in this configuration, the data mapping
is taken from the CH0, instead of the CH1 as should be.
To workarround this problem, when the mode selected is dual
or split, we configure both channels with the data mapping
passed through the command line.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit db349fbadeff2c94856f2a161e386b1a0f09475e)

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>