ccxmx53: da9052 battery, stop kthread on SIGKILL.
The SIGKILL signal is received on system poweroff/reboot. If the thread
is not stopped, the schedule() gets interrupted and the thread executes
several consecutive times. It also fails to access the I2C bus printing
error messages on the console.

Making the thread TASK_UNINTERRUPTIBLE before schedule stops the multiple
executions, but still prints I2C access error messages sometimes, when
the thread is scheduled in the middle of the poweroff/reboot.

Stopping the thread on SIGKILL provides a clean shutdown. The thread
won't stop on suspend as a SIGKILL is not received.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>