ccxmx53: Dialog I2C, perform a flush write on all accesses
There is a Dialog bug on the PMIC which can leave the I2C bus inaccessible
leaving an unbootable system.

The workaround suggested by Dialog is to flush all accesses with a last
write to a safe register.

They provided a patch which when applied does not work well, so this is a
rework simplified version of that patch.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>