ccxmx51: mc13892 battery, Fix driver to work without battery.
The mc13892 driver from the BSP assumed that the battery was always
connected. Otherwise the battery driver would not be loaded.

This fix allows for the driver to work even when booting without a
battery. It modifies the battery status function to use the charger
current measured via the ADC to detect the presence of a battery and
its charging status instead of the sense bit. Also, because now the
driver can tell whether a battery is present or not, it fixes some of
the places where the driver assumed the battery would always be connected.

This fixes 42033.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit 093ee454469036f25f104c2d544ba144d9a1ed37)