ccxmx53: add pmic gpio support to reset wifi (#41841)
This commit enables the support for the gpios of the da9053.
This allows to manupulate that gpios (for instance in the
wireless drivers, where the wifi power enable comes from a
pmic gpio).
Not doing a power cycling on the redpine module was causing
some detection issues on some ccwmx53 modules.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>