ccxmx5x: mma7455, minimize the possibility of an interrupted read.
The MMA7455 seems to have a hardware bug that makes it keep the data line
low if an I2C read transfer is interrupted between writing the register
address and reading the data.

For the CCWMX53 this means leaving the bus where the PMIC is inaccessible,
making it impossible to suspend or reboot, and needing a cold boot with
the coin cell disabled to recover.

Without a definitive fix for the hardware problem I have implemented a
patch to make the situation highly unlikely.

1) The register read is changed from two i2c transfers (one for writing
the register address and one for reading) to a single I2C transfer.
2) I have also added a 10ms delay between sysfs reads which helps
minimize the window of opportunity for the signal to interrupt the
I2C transfer.

These makes the problem very rare both for the MX51 and MX53.
I have not been able to repro it manually after the changes, but probably
an automated test could.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>