da9052: Perform a last PMIC access on poweroff.
Even though it should not be needed as the write_many on the same function
also "parks" the PMIC, meaning it does a last access to a safe register
to avoid a PMIC I2C bug, some modules need this last access to work.

Also, change the parking access to the STATUSA safe register, as the one
we had does not seem to be safe in all cases. That is the same register
used in WCE.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>