iomux mx53: set SION for user LEDs gpios
Test of idigibuttonled on ccwmx53js revealed that the status
of the user LEDs (gpios as outputs) always read as 0,
independently of their real status.
Investigation led to the SION (Software Input On Field) bit on
the MUX_CTL_PAD register. From the manual:

Software Input On Field. Force the selected mux mode Input path
no matter of MUX_MODE functionality.
 1 Force input path
 0 Input Path is determined by functionality of the selected
   mux mode (regular).

Default value of SION is 0, which makes the Linux driver always
reads 0 on the PSR (pad status register) of the gpio pins configured
as outputs, no matter the real value of the pin.

For the UserLEDs we want SION to be set, so that we can read the
real value of the pins.

This fixes vantive 42103.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>