ccxmx53 gpio: configure wifi SD interface as gpio on remove
This commit configures the SD pins connected to the wireless
module as gpios inputs when the driver is removed. This is
needed to avoid some kind of current feedback that happens
through the SD pins that causes that the reset level of the
redpine module doesnt go low enough causing module reload
problems (at least in WinCE).
Configuring the pins as inputs seems to solve the problem.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>