mtd: fsl_nfc: operate with 60-byte hardware ECC
Use 60-byte hardware ECC which allows up to 32-bit errors corrected
per 2k page. This is required for newer Macronix SLC NAND flashes.

Beside that, rename the driver to vf610_nfc, which is the new driver
name used in the upstream driver.

This makes this 3.18 kernel compatible with the new ECC format used
in the V2.4 Beta 1 release/U-Boot 2015.04.