irqchip: vf610-mscm-ir: alter local CPU's enable bit only
The driver clears the other CPU's interrupt enable bit, when enabling
a certain interrupt as well as when disabling an interrupt. This is
done intentionally, the idea behind it was to (mis)use the interrupt
router as kind of a resource manager. However, it turned out that
there are some interrupts which need to be assigned to both CPU, e.g.
the CPU2CPU interrupts.

Drop the poor man's resource manager functionality and only set/clear
the local CPU's enable bit.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>