mxc_hdmi: fix detect disconnect case
The keepalive module variable allows to not go into a disconnected
state. Fix the case when keepalive is set to go to disconnect.

Add a precompiler directive which allows to only evaluate the HPD signal
for HDMI monitor detection. Default is to react on the status bits for
HPD and the TMDS signal line detection to detect an attached monitor.

While at it cleanup parameter documentation and whitespace.