ARM: imx: busfreq: request bus frequency depending on M4 clock
Only request high bus frequency for Cortex-M4 if the Cortex-M4 is
running in a high frequency mode.  The Cortex-M4 clock is not
directly connected with the bus frequency, however, we assume that
when the M4 CPU is running at lower freuqency a lower bus speed is
sufficient too.

The driver already has been doing this on i.MX 6SoloX, this change
applies the same logic to i.MX 7.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>