ARM: dts: imx7-colibri: use pull-down for USBC_DET pin
The Standard Colibri pin for Client/Host detection USBC_DET
(SODIMM pin 137) is multiplexed on the module with USB_OTG1_VBUS.
The USB_OTG1_VBUS pad seems to be pulled low by default and not
controllable through pinctrl. When the pin is unconnected on the
Carrier board, the pin is at a intermediate voltage level which is
not ideal. Pull-down the pin to make sure that the pin is at a
defined state by default (low, USB host).

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>