tty: serial: fsl_lpuart: fix framing error handling when using DMA
When using DMA framing error get cleared properly. However, due
to the additional read from the data register, an underflow in
the receive FIFO buffer occures (the FIFO pointer gets out of
sync).

Clear the FIFO in case an underflow has occured. Also disable
the receiver during this operation and when reading the data
register to minimize potential interference.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>