imx8: isi: XXX: Add code to force format setting in the actual sensor
This commit adds code that forces setting the requested format in the
actual sensor driver.

Note that currently it is commented out as imx8 isi does not seem to
work properly with any other source resolution than the default
1920x1080.

Without this code, the driver will work with other resolutions (the ones
which are reported as supported by the connected sensors and lower than
1920x1080), but the image will be scaled down from 1920x1080. Scaling up
is not supported.